Move enums to a dedicated namespace
