+++ src/Draco/IO/Attributes/GeometryAttribute.cs
@@ -1,5 +1,6 @@
 using System.Numerics;
 using Draco.IO.Core;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Attributes;
