+++ src/Draco/IO/Attributes/PredictionSchemes/MeshPredictionSchemeConstrainedMultiParallelogramDecoder.cs
@@ -1,6 +1,7 @@
 using System.Numerics;
 using Draco.IO.BitCoders;
 using Draco.IO.Core;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Attributes.PredictionSchemes;
