+++ src/Draco/IO/Attributes/PredictionSchemes/MeshPredictionSchemeTexCoordsDecoder.cs
@@ -2,6 +2,7 @@ using System.Net.Http.Headers;
 using System.Numerics;
 using Draco.IO.BitCoders;
 using Draco.IO.Core;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Attributes.PredictionSchemes;
