+++ src/Draco/IO/Attributes/PredictionSchemes/MeshPredictionSchemeTexCoordsPortableDecoder.cs
@@ -1,5 +1,6 @@
 using System.Numerics;
 using Draco.IO.BitCoders;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Attributes.PredictionSchemes;
