+++ src/Draco/IO/Attributes/PredictionSchemes/PredictionSchemeDecoderFactory.cs
@@ -1,4 +1,5 @@
 using System.Numerics;
+using Draco.IO.Enums;
 using Draco.IO.Mesh;
 
 namespace Draco.IO.Attributes.PredictionSchemes;
