+++ src/Draco/IO/Attributes/PredictionSchemes/PredictionSchemeWrapTransform.cs
@@ -1,4 +1,5 @@
 using System.Numerics;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Attributes.PredictionSchemes;
