+++ src/Draco/IO/Attributes/SequentialIntegerAttributeDecoder.cs
@@ -1,5 +1,6 @@
 using Draco.IO.Attributes.PredictionSchemes;
 using Draco.IO.Entropy;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Attributes;
