+++ src/Draco/IO/Attributes/SequentialQuantizationAttributeDecoder.cs
@@ -1,3 +1,4 @@
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Attributes;
