+++ src/Draco/IO/Enums/DataType.cs
@@ -1,0 +1,18 @@
+namespace Draco.IO.Enums;
+
+public enum DataType : byte
+{
+    Invalid = 0,
+    Int8,
+    UInt8,
+    Int16,
+    UInt16,
+    Int32,
+    UInt32,
+    Int64,
+    UInt64,
+    Float32,
+    Float64,
+    Bool,
+    Count
+}
