+++ src/Draco/IO/Enums/GeometryAttributeType.cs
@@ -1,0 +1,12 @@
+namespace Draco.IO.Enums;
+
+public enum GeometryAttributeType : sbyte
+{
+    Invalid = -1,
+    Position = 0,
+    Normal = 1,
+    Color = 2,
+    TexCoord = 3,
+    Generic,
+    NamedAttributesCount
+}
