+++ src/Draco/IO/Enums/MeshTraversalMethod.cs
@@ -1,0 +1,8 @@
+namespace Draco.IO.Enums;
+
+public enum MeshTraversalMethod : byte
+{
+    DepthFirst = 0,
+    PredictionDegree,
+    Count
+}
