+++ src/Draco/IO/Enums/MultiParallelogramMode.cs
@@ -1,0 +1,6 @@
+namespace Draco.IO.Enums;
+
+public enum MultiParallelogramMode : byte
+{
+    Optimal = 0
+}
