+++ src/Draco/IO/Enums/NormalPredictionMode.cs
@@ -1,0 +1,7 @@
+namespace Draco.IO.Enums;
+
+public enum NormalPredictionMode : byte
+{
+    OneTriangle = 0,
+    TriangleArea = 1,
+}
