+++ src/Draco/IO/Enums/PredictionSchemeTransformType.cs
@@ -1,0 +1,11 @@
+namespace Draco.IO.Enums;
+
+public enum PredictionSchemeTransformType : sbyte
+{
+    None = -1,
+    Delta = 0,
+    Wrap = 1,
+    NormalOctahedron = 2,
+    NormalOctahedronCanonicalized = 3,
+    Count,
+}
