+++ src/Draco/IO/Enums/SequentialAttributeEncoderType.cs
@@ -1,0 +1,9 @@
+namespace Draco.IO.Enums;
+
+public enum SequentialAttributeEncoderType : byte
+{
+    Generic = 0,
+    Integer,
+    Quantization,
+    Normals
+}
