+++ src/Draco/IO/Mesh/Mesh.cs
@@ -1,4 +1,5 @@
 using Draco.IO.Attributes;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 
 namespace Draco.IO.Mesh;
