+++ src/Draco/IO/Mesh/MeshEdgeBreakerDecoder.cs
@@ -1,4 +1,5 @@
 using Draco.IO.Attributes;
+using Draco.IO.Enums;
 using Draco.IO.Extensions;
 using Draco.IO.Mesh.Traverser;
